rmp-ai-assistant-python/requirements.txt
pinecone-client
openai
python-dotenv